subclass nlp_keras transformer encoder for deit
