+++ official/vision/beta/projects/vit/modeling/layers/__init__.py
@@ -1,1 +1,1 @@
-from official.vision.beta.projects.vit.modeling.layers.vit_transformer_encoder_block import TransformerEncoderBlock
+from official.vision.beta.projects.vit.modeling.layers.transformer_encoder_block import StochsticDepthTransformerEncoderBlock
